+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/AssetDataFactory.java
@@ -1,0 +1,631 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.tools.sample.sql.builder;
+
+import com.liferay.asset.kernel.model.AssetCategoryConstants;
+import com.liferay.asset.kernel.model.AssetCategoryModel;
+import com.liferay.asset.kernel.model.AssetEntry;
+import com.liferay.asset.kernel.model.AssetEntryModel;
+import com.liferay.asset.kernel.model.AssetTagModel;
+import com.liferay.asset.kernel.model.AssetVocabularyModel;
+import com.liferay.blogs.model.BlogsEntry;
+import com.liferay.blogs.model.BlogsEntryModel;
+import com.liferay.commerce.product.model.CPDefinition;
+import com.liferay.document.library.kernel.model.DLFileEntry;
+import com.liferay.document.library.kernel.model.DLFileEntryModel;
+import com.liferay.document.library.kernel.model.DLFolder;
+import com.liferay.document.library.kernel.model.DLFolderModel;
+import com.liferay.journal.model.JournalArticle;
+import com.liferay.journal.model.JournalArticleLocalizationModel;
+import com.liferay.journal.model.JournalArticleModel;
+import com.liferay.message.boards.constants.MBCategoryConstants;
+import com.liferay.message.boards.model.MBDiscussion;
+import com.liferay.message.boards.model.MBMessage;
+import com.liferay.message.boards.model.MBMessageModel;
+import com.liferay.message.boards.model.MBThread;
+import com.liferay.message.boards.model.MBThreadModel;
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.util.ContentTypes;
+import com.liferay.portal.kernel.util.ObjectValuePair;
+import com.liferay.portal.kernel.util.StringBundler;
+import com.liferay.portlet.asset.model.impl.AssetCategoryModelImpl;
+import com.liferay.portlet.asset.model.impl.AssetEntryModelImpl;
+import com.liferay.portlet.asset.model.impl.AssetTagModelImpl;
+import com.liferay.portlet.asset.model.impl.AssetVocabularyModelImpl;
+import com.liferay.util.SimpleCounter;
+import com.liferay.view.count.model.ViewCountEntryModel;
+import com.liferay.view.count.model.impl.ViewCountEntryModelImpl;
+import com.liferay.view.count.service.persistence.ViewCountEntryPK;
+import com.liferay.wiki.model.WikiPage;
+import com.liferay.wiki.model.WikiPageModel;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+/**
+ * @author Lily Chi
+ */
+public class AssetDataFactory extends BaseDataFactory {
+
+	public AssetDataFactory() {
+		_initAssetCategoryModels();
+		_initAssetTagModels();
+	}
+
+	public List<Long> getAssetCategoryIds(AssetEntryModel assetEntryModel) {
+		Map<Long, List<AssetCategoryModel>> assetCategoryModelsMap =
+			assetCategoryModelsMaps[(int)assetEntryModel.getGroupId() - 1];
+
+		if ((assetCategoryModelsMap == null) ||
+			assetCategoryModelsMap.isEmpty()) {
+
+			return Collections.emptyList();
+		}
+
+		List<AssetCategoryModel> assetCategoryModels =
+			assetCategoryModelsMap.get(assetEntryModel.getClassNameId());
+
+		if ((assetCategoryModels == null) || assetCategoryModels.isEmpty()) {
+			return Collections.emptyList();
+		}
+
+		if (_assetCategoryCounters == null) {
+			_assetCategoryCounters =
+				(Map<Long, SimpleCounter>[])
+					new HashMap<?, ?>[PropsValues.MAX_GROUP_COUNT];
+		}
+
+		SimpleCounter counter = getSimpleCounter(
+			_assetCategoryCounters, assetEntryModel.getGroupId(),
+			assetEntryModel.getClassNameId());
+
+		List<Long> assetCategoryIds = new ArrayList<>(
+			PropsValues.MAX_ASSET_ENTRY_TO_ASSET_CATEGORY_COUNT);
+
+		for (int i = 0; i < PropsValues.MAX_ASSET_ENTRY_TO_ASSET_CATEGORY_COUNT;
+			 i++) {
+
+			int index = (int)counter.get() % assetCategoryModels.size();
+
+			AssetCategoryModel assetCategoryModel = assetCategoryModels.get(
+				index);
+
+			assetCategoryIds.add(assetCategoryModel.getCategoryId());
+		}
+
+		return assetCategoryIds;
+	}
+
+	public List<AssetCategoryModel> getAssetCategoryModels() {
+		List<AssetCategoryModel> allAssetCategoryModels = new ArrayList<>();
+
+		for (Map<Long, List<AssetCategoryModel>> assetCategoryModelsMap :
+				assetCategoryModelsMaps) {
+
+			for (List<AssetCategoryModel> assetCategoryModels :
+					assetCategoryModelsMap.values()) {
+
+				allAssetCategoryModels.addAll(assetCategoryModels);
+			}
+		}
+
+		return allAssetCategoryModels;
+	}
+
+	public List<Long> getAssetTagIds(AssetEntryModel assetEntryModel) {
+		Map<Long, List<AssetTagModel>> assetTagModelsMap =
+			assetTagModelsMaps[(int)assetEntryModel.getGroupId() - 1];
+
+		if ((assetTagModelsMap == null) || assetTagModelsMap.isEmpty()) {
+			return Collections.emptyList();
+		}
+
+		List<AssetTagModel> assetTagModels = assetTagModelsMap.get(
+			assetEntryModel.getClassNameId());
+
+		if ((assetTagModels == null) || assetTagModels.isEmpty()) {
+			return Collections.emptyList();
+		}
+
+		if (_assetTagCounters == null) {
+			_assetTagCounters =
+				(Map<Long, SimpleCounter>[])
+					new HashMap<?, ?>[PropsValues.MAX_GROUP_COUNT];
+		}
+
+		SimpleCounter counter = getSimpleCounter(
+			_assetTagCounters, assetEntryModel.getGroupId(),
+			assetEntryModel.getClassNameId());
+
+		List<Long> assetTagIds = new ArrayList<>(
+			PropsValues.MAX_ASSET_ENTRY_TO_ASSET_TAG_COUNT);
+
+		for (int i = 0; i < PropsValues.MAX_ASSET_ENTRY_TO_ASSET_TAG_COUNT;
+			 i++) {
+
+			int index = (int)counter.get() % assetTagModels.size();
+
+			AssetTagModel assetTagModel = assetTagModels.get(index);
+
+			assetTagIds.add(assetTagModel.getTagId());
+		}
+
+		return assetTagIds;
+	}
+
+	public List<AssetTagModel> getAssetTagModels() {
+		List<AssetTagModel> allAssetTagModels = new ArrayList<>();
+
+		for (Map<Long, List<AssetTagModel>> assetTagModelsMap :
+				assetTagModelsMaps) {
+
+			for (List<AssetTagModel> assetTagModels :
+					assetTagModelsMap.values()) {
+
+				allAssetTagModels.addAll(assetTagModels);
+			}
+		}
+
+		return allAssetTagModels;
+	}
+
+	public List<AssetVocabularyModel> getAssetVocabularyModels() {
+		List<AssetVocabularyModel> allAssetVocabularyModels = new ArrayList<>();
+
+		allAssetVocabularyModels.add(_defaultAssetVocabularyModel);
+
+		for (List<AssetVocabularyModel> assetVocabularyModels :
+				_assetVocabularyModelsArray) {
+
+			allAssetVocabularyModels.addAll(assetVocabularyModels);
+		}
+
+		return allAssetVocabularyModels;
+	}
+
+	public int getMaxAssetPublisherPageCount() {
+		return PropsValues.MAX_ASSETPUBLISHER_PAGE_COUNT;
+	}
+
+	public AssetEntryModel newAssetEntryModel(BlogsEntryModel blogsEntryModel) {
+		return newAssetEntryModel(
+			blogsEntryModel.getGroupId(), blogsEntryModel.getCreateDate(),
+			blogsEntryModel.getModifiedDate(), getClassNameId(BlogsEntry.class),
+			blogsEntryModel.getEntryId(), blogsEntryModel.getUuid(), 0, true,
+			true, ContentTypes.TEXT_HTML, blogsEntryModel.getTitle());
+	}
+
+	public AssetEntryModel newAssetEntryModel(
+		DLFileEntryModel dLFileEntryModel) {
+
+		return newAssetEntryModel(
+			dLFileEntryModel.getGroupId(), dLFileEntryModel.getCreateDate(),
+			dLFileEntryModel.getModifiedDate(),
+			getClassNameId(DLFileEntry.class),
+			dLFileEntryModel.getFileEntryId(), dLFileEntryModel.getUuid(),
+			dLFileEntryModel.getFileEntryTypeId(), true, true,
+			dLFileEntryModel.getMimeType(), dLFileEntryModel.getTitle());
+	}
+
+	public AssetEntryModel newAssetEntryModel(DLFolderModel dLFolderModel) {
+		return newAssetEntryModel(
+			dLFolderModel.getGroupId(), dLFolderModel.getCreateDate(),
+			dLFolderModel.getModifiedDate(), getClassNameId(DLFolder.class),
+			dLFolderModel.getFolderId(), dLFolderModel.getUuid(), 0, true, true,
+			null, dLFolderModel.getName());
+	}
+
+	public AssetEntryModel newAssetEntryModel(MBMessageModel mbMessageModel) {
+		long classNameId = 0;
+		boolean visible = false;
+
+		if (mbMessageModel.getCategoryId() ==
+				MBCategoryConstants.DISCUSSION_CATEGORY_ID) {
+
+			classNameId = getClassNameId(MBDiscussion.class);
+		}
+		else {
+			classNameId = getClassNameId(MBMessage.class);
+			visible = true;
+		}
+
+		return newAssetEntryModel(
+			mbMessageModel.getGroupId(), mbMessageModel.getCreateDate(),
+			mbMessageModel.getModifiedDate(), classNameId,
+			mbMessageModel.getMessageId(), mbMessageModel.getUuid(), 0, true,
+			visible, ContentTypes.TEXT_HTML, mbMessageModel.getSubject());
+	}
+
+	public AssetEntryModel newAssetEntryModel(MBThreadModel mbThreadModel) {
+		return newAssetEntryModel(
+			mbThreadModel.getGroupId(), mbThreadModel.getCreateDate(),
+			mbThreadModel.getModifiedDate(), getClassNameId(MBThread.class),
+			mbThreadModel.getThreadId(), mbThreadModel.getUuid(), 0, true,
+			false, StringPool.BLANK,
+			String.valueOf(mbThreadModel.getRootMessageId()));
+	}
+
+	public AssetEntryModel newAssetEntryModel(
+		ObjectValuePair<JournalArticleModel, JournalArticleLocalizationModel>
+			objectValuePair) {
+
+		JournalArticleModel journalArticleModel = objectValuePair.getKey();
+		JournalArticleLocalizationModel journalArticleLocalizationModel =
+			objectValuePair.getValue();
+
+		long resourcePrimKey = journalArticleModel.getResourcePrimKey();
+
+		String resourceUUID = journalArticleResourceUUIDs.get(resourcePrimKey);
+
+		return newAssetEntryModel(
+			journalArticleModel.getGroupId(),
+			journalArticleModel.getCreateDate(),
+			journalArticleModel.getModifiedDate(),
+			getClassNameId(JournalArticle.class), resourcePrimKey, resourceUUID,
+			DEFAULT_JOURNAL_DDM_STRUCTURE_ID, journalArticleModel.isIndexable(),
+			true, ContentTypes.TEXT_HTML,
+			journalArticleLocalizationModel.getTitle());
+	}
+
+	public AssetEntryModel newAssetEntryModel(WikiPageModel wikiPageModel) {
+		return newAssetEntryModel(
+			wikiPageModel.getGroupId(), wikiPageModel.getCreateDate(),
+			wikiPageModel.getModifiedDate(), getClassNameId(WikiPage.class),
+			wikiPageModel.getResourcePrimKey(), wikiPageModel.getUuid(), 0,
+			true, true, ContentTypes.TEXT_HTML, wikiPageModel.getTitle());
+	}
+
+	public List<AssetEntryModel> newAssetEntryModels() {
+		List<AssetEntryModel> assetEntryModels = new ArrayList<>(
+			CPDEFINITION_COUNT);
+
+		for (int productIndex = 0;
+			 productIndex < PropsValues.MAX_COMMERCE_PRODUCT_COUNT;
+			 productIndex++) {
+
+			long[] cpDefinitionIds = (long[])cpDefinitionIdList.get(
+				productIndex);
+
+			for (int definitionIndex = 0;
+				 definitionIndex <
+					 PropsValues.MAX_COMMERCE_PRODUCT_DEFINITION_COUNT;
+				 definitionIndex++) {
+
+				assetEntryModels.add(
+					newAssetEntryModel(
+						COMMERCE_CATALOG_GROUP_ID, new Date(), new Date(),
+						getClassNameId(CPDefinition.class),
+						cpDefinitionIds[definitionIndex],
+						SequentialUUID.generate(), 0, true, true, "text/plain",
+						cpDefinitionLocalizationNames.get(
+							cpDefinitionIds[definitionIndex])));
+			}
+		}
+
+		return assetEntryModels;
+	}
+
+	public AssetEntryModel newMBDiscussionAssetEntryModel(
+		BlogsEntryModel blogsEntryModel) {
+
+		return newAssetEntryModel(
+			blogsEntryModel.getGroupId(), blogsEntryModel.getCreateDate(),
+			blogsEntryModel.getModifiedDate(),
+			getClassNameId(getMBDiscussionCombinedClassName(BlogsEntry.class)),
+			blogsEntryModel.getEntryId(), "", 0, true, false, "",
+			String.valueOf(blogsEntryModel.getGroupId()));
+	}
+
+	public AssetEntryModel newMBDiscussionAssetEntryModel(
+		WikiPageModel wikiPageModel) {
+
+		return newAssetEntryModel(
+			wikiPageModel.getGroupId(), wikiPageModel.getCreateDate(),
+			wikiPageModel.getModifiedDate(),
+			getClassNameId(getMBDiscussionCombinedClassName(WikiPage.class)),
+			wikiPageModel.getResourcePrimKey(), "", 0, true, false, "",
+			String.valueOf(wikiPageModel.getGroupId()));
+	}
+
+	public ViewCountEntryModel newViewCountEntryModel(
+		AssetEntryModel assetEntryModel) {
+
+		return newViewCountEntryModel(
+			assetEntryModel.getCompanyId(), getClassNameId(AssetEntry.class),
+			assetEntryModel.getPrimaryKey(), 0);
+	}
+
+	protected SimpleCounter getSimpleCounter(
+		Map<Long, SimpleCounter>[] simpleCountersArray, long groupId,
+		long classNameId) {
+
+		Map<Long, SimpleCounter> simpleCounters =
+			simpleCountersArray[(int)groupId - 1];
+
+		if (simpleCounters == null) {
+			simpleCounters = new HashMap<>();
+
+			simpleCountersArray[(int)groupId - 1] = simpleCounters;
+		}
+
+		SimpleCounter simpleCounter = simpleCounters.get(classNameId);
+
+		if (simpleCounter == null) {
+			simpleCounter = new SimpleCounter(0);
+
+			simpleCounters.put(classNameId, simpleCounter);
+		}
+
+		return simpleCounter;
+	}
+
+	protected AssetCategoryModel newAssetCategoryModel(
+		long groupId, String name, long vocabularyId) {
+
+		AssetCategoryModel assetCategoryModel = new AssetCategoryModelImpl();
+
+		assetCategoryModel.setUuid(SequentialUUID.generate());
+		assetCategoryModel.setCategoryId(counter.get());
+		assetCategoryModel.setGroupId(groupId);
+		assetCategoryModel.setCompanyId(COMPANY_ID);
+		assetCategoryModel.setUserId(SAMPLE_USER_ID);
+		assetCategoryModel.setUserName(DataFactoryConstants.SAMPLE_USER_NAME);
+		assetCategoryModel.setCreateDate(new Date());
+		assetCategoryModel.setModifiedDate(new Date());
+		assetCategoryModel.setParentCategoryId(
+			AssetCategoryConstants.DEFAULT_PARENT_CATEGORY_ID);
+		assetCategoryModel.setTreePath(
+			"/" + assetCategoryModel.getCategoryId() + "/");
+		assetCategoryModel.setName(name);
+
+		StringBundler sb = new StringBundler(4);
+
+		sb.append("<?xml version=\"1.0\"?><root available-locales=\"en_US\" ");
+		sb.append("default-locale=\"en_US\"><Title language-id=\"en_US\">");
+		sb.append(name);
+		sb.append("</Title></root>");
+
+		assetCategoryModel.setTitle(sb.toString());
+
+		assetCategoryModel.setVocabularyId(vocabularyId);
+		assetCategoryModel.setLastPublishDate(new Date());
+
+		return assetCategoryModel;
+	}
+
+	protected AssetEntryModel newAssetEntryModel(
+		long groupId, Date createDate, Date modifiedDate, long classNameId,
+		long classPK, String uuid, long classTypeId, boolean listable,
+		boolean visible, String mimeType, String title) {
+
+		AssetEntryModel assetEntryModel = new AssetEntryModelImpl();
+
+		assetEntryModel.setEntryId(counter.get());
+		assetEntryModel.setGroupId(groupId);
+		assetEntryModel.setCompanyId(COMPANY_ID);
+		assetEntryModel.setUserId(SAMPLE_USER_ID);
+		assetEntryModel.setUserName(DataFactoryConstants.SAMPLE_USER_NAME);
+		assetEntryModel.setCreateDate(createDate);
+		assetEntryModel.setModifiedDate(modifiedDate);
+		assetEntryModel.setClassNameId(classNameId);
+		assetEntryModel.setClassPK(classPK);
+		assetEntryModel.setClassUuid(uuid);
+		assetEntryModel.setClassTypeId(classTypeId);
+		assetEntryModel.setListable(listable);
+		assetEntryModel.setVisible(visible);
+		assetEntryModel.setStartDate(createDate);
+		assetEntryModel.setEndDate(nextFutureDate());
+		assetEntryModel.setPublishDate(createDate);
+		assetEntryModel.setExpirationDate(nextFutureDate());
+		assetEntryModel.setMimeType(mimeType);
+		assetEntryModel.setTitle(title);
+
+		return assetEntryModel;
+	}
+
+	protected AssetVocabularyModel newAssetVocabularyModel(
+		long grouId, long userId, String userName, String name) {
+
+		AssetVocabularyModel assetVocabularyModel =
+			new AssetVocabularyModelImpl();
+
+		assetVocabularyModel.setUuid(SequentialUUID.generate());
+		assetVocabularyModel.setVocabularyId(counter.get());
+		assetVocabularyModel.setGroupId(grouId);
+		assetVocabularyModel.setCompanyId(COMPANY_ID);
+		assetVocabularyModel.setUserId(userId);
+		assetVocabularyModel.setUserName(userName);
+		assetVocabularyModel.setCreateDate(new Date());
+		assetVocabularyModel.setModifiedDate(new Date());
+		assetVocabularyModel.setName(name);
+
+		StringBundler sb = new StringBundler(4);
+
+		sb.append("<?xml version=\"1.0\"?><root available-locales=\"en_US\" ");
+		sb.append("default-locale=\"en_US\"><Title language-id=\"en_US\">");
+		sb.append(name);
+		sb.append("</Title></root>");
+
+		assetVocabularyModel.setTitle(sb.toString());
+
+		assetVocabularyModel.setSettings(
+			"multiValued=true\\nselectedClassNameIds=0");
+		assetVocabularyModel.setLastPublishDate(new Date());
+
+		return assetVocabularyModel;
+	}
+
+	protected ViewCountEntryModel newViewCountEntryModel(
+		long companyId, long classNameId, long classPK, long viewCount) {
+
+		ViewCountEntryModel viewCountEntryModel = new ViewCountEntryModelImpl();
+
+		viewCountEntryModel.setPrimaryKey(new ViewCountEntryPK());
+		viewCountEntryModel.setCompanyId(companyId);
+		viewCountEntryModel.setClassNameId(classNameId);
+		viewCountEntryModel.setClassPK(classPK);
+		viewCountEntryModel.setViewCount(viewCount);
+
+		return viewCountEntryModel;
+	}
+
+	private void _initAssetCategoryModels() {
+		_assetCategoryModelsArray =
+			(List<AssetCategoryModel>[])
+				new List<?>[PropsValues.MAX_GROUP_COUNT];
+		assetCategoryModelsMaps =
+			(Map<Long, List<AssetCategoryModel>>[])
+				new HashMap<?, ?>[PropsValues.MAX_GROUP_COUNT];
+		_assetVocabularyModelsArray =
+			(List<AssetVocabularyModel>[])
+				new List<?>[PropsValues.MAX_GROUP_COUNT];
+		_defaultAssetVocabularyModel = newAssetVocabularyModel(
+			GLOBAL_GROUP_ID, DEFAULT_USER_ID, null,
+			com.liferay.portal.util.PropsValues.ASSET_VOCABULARY_DEFAULT);
+
+		StringBundler sb = new StringBundler(4);
+
+		for (int i = 1; i <= PropsValues.MAX_GROUP_COUNT; i++) {
+			List<AssetVocabularyModel> assetVocabularyModels = new ArrayList<>(
+				PropsValues.MAX_ASSET_VUCABULARY_COUNT);
+			List<AssetCategoryModel> assetCategoryModels = new ArrayList<>(
+				PropsValues.MAX_ASSET_VUCABULARY_COUNT *
+					PropsValues.MAX_ASSET_CATEGORY_COUNT);
+
+			for (int j = 0; j < PropsValues.MAX_ASSET_VUCABULARY_COUNT; j++) {
+				sb.setIndex(0);
+
+				sb.append(DataFactoryConstants.ASSET_VOCABULARY_NAME_PREFIX);
+				sb.append(i);
+				sb.append(StringPool.UNDERLINE);
+				sb.append(j);
+
+				AssetVocabularyModel assetVocabularyModel =
+					newAssetVocabularyModel(
+						i, SAMPLE_USER_ID,
+						DataFactoryConstants.SAMPLE_USER_NAME, sb.toString());
+
+				assetVocabularyModels.add(assetVocabularyModel);
+
+				for (int k = 0; k < PropsValues.MAX_ASSET_CATEGORY_COUNT; k++) {
+					sb.setIndex(0);
+
+					sb.append(DataFactoryConstants.ASSET_CATEGORY_NAME_PREFIX);
+					sb.append(assetVocabularyModel.getVocabularyId());
+					sb.append(StringPool.UNDERLINE);
+					sb.append(k);
+
+					assetCategoryModels.add(
+						newAssetCategoryModel(
+							i, sb.toString(),
+							assetVocabularyModel.getVocabularyId()));
+				}
+			}
+
+			_assetCategoryModelsArray[i - 1] = assetCategoryModels;
+			_assetVocabularyModelsArray[i - 1] = assetVocabularyModels;
+
+			Map<Long, List<AssetCategoryModel>> assetCategoryModelsMap =
+				new HashMap<>();
+
+			int pageSize =
+				assetCategoryModels.size() / assetClassNameIds.length;
+
+			for (int j = 0; j < assetClassNameIds.length; j++) {
+				int fromIndex = j * pageSize;
+
+				int toIndex = (j + 1) * pageSize;
+
+				if (j == (assetClassNameIds.length - 1)) {
+					toIndex = assetCategoryModels.size();
+				}
+
+				assetCategoryModelsMap.put(
+					assetClassNameIds[j],
+					assetCategoryModels.subList(fromIndex, toIndex));
+			}
+
+			assetCategoryModelsMaps[i - 1] = assetCategoryModelsMap;
+		}
+	}
+
+	private void _initAssetTagModels() {
+		_assetTagModelsArray =
+			(List<AssetTagModel>[])new List<?>[PropsValues.MAX_GROUP_COUNT];
+		assetTagModelsMaps =
+			(Map<Long, List<AssetTagModel>>[])
+				new HashMap<?, ?>[PropsValues.MAX_GROUP_COUNT];
+
+		for (int i = 1; i <= PropsValues.MAX_GROUP_COUNT; i++) {
+			List<AssetTagModel> assetTagModels = new ArrayList<>(
+				PropsValues.MAX_ASSET_TAG_COUNT);
+
+			for (int j = 0; j < PropsValues.MAX_ASSET_TAG_COUNT; j++) {
+				AssetTagModel assetTagModel = new AssetTagModelImpl();
+
+				assetTagModel.setUuid(SequentialUUID.generate());
+				assetTagModel.setTagId(counter.get());
+				assetTagModel.setGroupId(i);
+				assetTagModel.setCompanyId(COMPANY_ID);
+				assetTagModel.setUserId(SAMPLE_USER_ID);
+				assetTagModel.setUserName(
+					DataFactoryConstants.SAMPLE_USER_NAME);
+				assetTagModel.setCreateDate(new Date());
+				assetTagModel.setModifiedDate(new Date());
+				assetTagModel.setName(
+					StringBundler.concat(
+						DataFactoryConstants.ASSET_TAG_NAME_PREFIX,
+						String.valueOf(i), "_", String.valueOf(j)));
+				assetTagModel.setLastPublishDate(new Date());
+
+				assetTagModels.add(assetTagModel);
+			}
+
+			_assetTagModelsArray[i - 1] = assetTagModels;
+
+			Map<Long, List<AssetTagModel>> assetTagModelsMap = new HashMap<>();
+
+			int pageSize = assetTagModels.size() / assetClassNameIds.length;
+
+			for (int j = 0; j < assetClassNameIds.length; j++) {
+				int fromIndex = j * pageSize;
+
+				int toIndex = (j + 1) * pageSize;
+
+				if (j == (assetClassNameIds.length - 1)) {
+					toIndex = assetTagModels.size();
+				}
+
+				assetTagModelsMap.put(
+					assetClassNameIds[j],
+					assetTagModels.subList(fromIndex, toIndex));
+			}
+
+			assetTagModelsMaps[i - 1] = assetTagModelsMap;
+		}
+	}
+
+	private Map<Long, SimpleCounter>[] _assetCategoryCounters;
+	private List<AssetCategoryModel>[] _assetCategoryModelsArray;
+	private Map<Long, SimpleCounter>[] _assetTagCounters;
+	private List<AssetTagModel>[] _assetTagModelsArray;
+	private List<AssetVocabularyModel>[] _assetVocabularyModelsArray;
+	private AssetVocabularyModel _defaultAssetVocabularyModel;
+
+}
