+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/commerce_product.ftl
@@ -1,4 +1,4 @@
-<#list dataFactory.newAssetEntryModels() as assetEntryModel>
+<#list assetDataFactory.newAssetEntryModels() as assetEntryModel>
 	${resourcePermissionDataFactory.toInsertSQL(assetEntryModel)}
 </#list>
